Ajout de la clé étrangère de la table `flux_batches` aux migrations.

+++ database/migrations/2019_03_13_130933_create_flux_batches_table.php
@@ -15,7 +15,7 @@ class CreateFluxBatchesTable extends Migration
     {
         Schema::create('flux_batches', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('flux_id');
+            $table->unsignedBigInteger('flux_id');
             $table->dateTime('started_at');
             $table->dateTime('ended_at');
             $table->boolean('success');
